requirements.txt
allure-pytest==2.13.5
selenium==4.20.0
pytest==8.2.0
